+++ test/nosedrum/converters/channel_test.exs
@@ -1,7 +1,7 @@
 defmodule Nosedrum.Converters.ChannelTest do
   alias Nosedrum.Converters.Channel, as: ChannelConverter
   alias Nostrum.Cache.CacheSupervisor
-  alias Nostrum.Cache.Guild.GuildRegister
+  alias Nostrum.Cache.GuildCache.ETS, as: GuildCache
   alias Nostrum.Struct.Channel
   alias Nostrum.Struct.Guild
   use ExUnit.Case
@@ -21,7 +21,7 @@ defmodule Nosedrum.Converters.ChannelTest do
       channels: [channel]
     }
 
-    GuildRegister.create_guild_process(guild.id, guild)
+    GuildCache.create(guild)
     %{channel: channel, guild: guild}
   end